refactor: Create type aliases for all database IDs (#98)
Change existing IDs to u32 to avoid issues when loading from SQLite

+++ src/tag/domain.rs
@@ -57,7 +57,7 @@ impl Display for TagName {
 }
 
 /// Database identifier for a tag.
-pub type TagId = i64;
+pub type TagId = u32;
 
 /// A tag for categorizing transactions (e.g., 'Groceries', 'Salary').
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Hash)]
